backend/requirements.txt
langchain>=0.3.27
langchain-openai>=0.3.28
langchain-community>=0.3.27
langchain-chroma>=0.2.5
langgraph>=0.6.0
langchainhub>=0.1.21
duckduckgo-search>=8.1.1
langchain-google-genai>=2.1.8
langchain-mcp-adapters>=0.1.9
langchain-perplexity>=0.1.2
langchain-tavily>=0.2.11
python-dotenv>=1.1.1
pandas>=2.3.1
matplotlib>=3.10.3
seaborn>=0.13.2
scikit-learn>=1.7.1
